+++ applovin_max/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Versions
 
+##
+* Update preloaded banners and MRECs (`MaxAdView`) to suspend auto-refresh while not visible in background.
 ## 4.0.2
 * Update IconView to support native ad icon image view, primarily for BigoAds native ads.
 ## 4.0.1