+++ applovin_max/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## x.x.x
 * Enhance banner and MREC (`MaxAdView`) preloading to support preloading multiple `MaxAdView` instances.
 * Update preloaded banners and MRECs (`MaxAdView`) to suspend auto-refresh while not visible in background.
+* Depends on Android SDK v13.0.1 and iOS SDK v13.0.1.
 ## 4.0.2
 * Update IconView to support native ad icon image view, primarily for BigoAds native ads.
 ## 4.0.1